# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Suite Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get full user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/history:
    get:
      operationId: getUserHistory
      tags: [user]
      summary: Get user generation history
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Generation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GenerationRecord"
...
  /user/favorites/{toolId}:
    post:
      operationId: addFavorite
      tags: [user]
      summary: Add tool to favorites
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeFavorite
      tags: [user]
      summary: Remove tool from favorites
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed from favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Plans ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── Auth ─────────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, email, name, role, tokenBalance, planId, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        tokenBalance:
          type: integer
        planId:
          type: ["string", "null"]
        planName:
          type: ["string", "null"]
        createdAt:
          type: string
        totalGenerations:
          type: integer
        favoriteToolIds:
          type: array
          items:
            type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    GenerationRecord:
      type: object
      required: [id, toolId, toolLabel, tokensUsed, createdAt]
      properties:
        id:
          type: integer
        toolId:
          type: string
        toolLabel:
          type: string
        prompt:
          type: ["string", "null"]
        output:
          type: ["string", "null"]
        tokensUsed:
          type: integer
        model:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Plans ────────────────────────────────────────────────────────────────
...