# lib/api-zod/src/generated/types/generationRecord.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Suite Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerationRecord {
  id: number;
  toolId: string;
  toolLabel: string;
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  output?: string | null;
  tokensUsed: number;
  /** @nullable */
  model?: string | null;
  createdAt: string;
}